Simplify algorithm for working out name and badge in one instance.
This commit modifies how 'name' and 'badge' are retrieved in the global
endpoint template by taking advantage of the fact that the configuration offers
a dictionary lookup interface. New functionality is intended to be
indistinguishable in all respects expect one: a missing 'name' attribute will
now trigger a user-readable error. Previously this likely caused an incorrect
`pop` from an empty list.