Remove type juggling of client_id in seL4MultiSharedData.
The templates for this connector converted `client_id` back and forth between a
string and an integer. The purpose of this seemed to be to validate that this
parameter was in fact a number. Following the previous commit, this is checked
more rigorously and provided to the user as an actionable error. The current
commit removes this juggling entirely.