templates: Use USED instead of keep_symbol
Keep symbol results in -Wl,-u symbol being passed to the linker, which
is equivalent to attribute((used)) that USED expands to.  Note: This
assumes that every generated file is passed to the compiler directly and
not within a library. In this case it may be required for -u to be used.