SerialServer: Change plat_pre_init to take IO ops
We change plat_pre_init to take an IO ops structure instead of
initliasing one because of the addition of an IRQ ops structure in the
IO ops structure. The IRQ ops structure is used in other parts of the
component and we do not want to leak it with the IO ops structure.