components: Add new GPIOMUXServer
This commit adds in a new CAmkES component: GPIOMUXServer. The component
is intended to multiplex the MUX and GPIO functionalities of a platform.
The component also includes a client library which abstracts over
libplatsupport/libplatsupportports and exposes the necessary constants
needed to interact with the component. At the moment, the server has
pretty basic functionality and only has exynos5 and tx2 support.