SerialServer: Fix fsm issue
When an invalid command is sent, the serial server attempted to give
the client the characters. By sending more than one character at once
the queue seemed to get messed up. Therefore, the serial interface
became almost impossible to use.

This change fixes that by assuming any command was meant to be a
command. Instead of sending the characters to the client, it prints an
error message and suggests using the help command. Then it throws away
the extra characters.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>