components: Add ClockServer component
The ClockServer component is a new component that is responsible for
managing and multiplexing requests to the clock controllers of a
particular platform. The functionality of the component is very limited
at the moment and can only support the functionality that is exposed via
the clock subsystem in libplatsupport. There is no access control
with regards to clock gates so components can interfere with other
components by manipulating the clock gates.