seL4SharedDataWithCaps: Correctly size data port
If a dataport is declared with a C type that is smaller than the
declared dataport size, the backing char[] symbol that is used to place
the dataport needs to be rounded up to the size of the dataport.
Additionally if the dataport is smaller than the symbol we want to catch
this with a static compile assert.