SerialServer: Fix switching instability
If the serial server received two \n and/or \r characters in a row,
then it would reset back to the beginning of the FSM. This resulted
in needing a certain timing to successfully switch VMs.

This change fixes this behavior by staying in the same state when
multiple \n and/or \r characters are received.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>