RISCV: Created riscv_common Serial Server plat
This enables the Serial Server component for the riscv_common
platform. These files were initially copied from the arm_common platform
files. The PlatPrefix is set based whether or not the KernelArchRiscV
variable is set. This also requires that the device tree for the
platform has a valid "chosen" serial "stdout-path" node.

Signed-off-by: Jesse Millwood <jesse.millwood@dornerworks.com>