templates: Standardize use of assigning badges
Don't require ${interface_name}_attributes to be set in every case. The
'from' ends don't typically need to control what badge they appear under
in the 'to' end and it is easier to allow the connector to allocate
badges itself.  Note: it may still be required to force badge assignment
in order to link separate interfaces together and so this mechanism is
still supported.