seL4SharedDataWithCaps: Fix dataport size config
Previously the dataport in this connector was sized by setting a
configuration on the interface on the `to` end. A previous commit
mistakenly read the configuration on either end but would require both
ends to be set correctly. This commit reverts to the old behavior but
also adds a secondary method which is the same as seL4SharedData where
the size can be set as a configuration on the connection itself.