plat,tx2: Add tx2-specific BPMPServer component
This commit adds a new component, BPMPServer, which is intended to
wrap over a BPMP interface and multiplex it to multiple clients. The
component also contains a client library which given the functions
exposed by the CAmkES procedure, wraps these inside a BPMP interface as
defined in libtx2bpmp. This makes it easier to pass to use in
CAmkES-agnostic libraries and code such as those in libplatsupport.