ClockServer,tx2: Add BPMP connector definitions
Instead of initialising the BPMP component alongside the ClockServer,
this commit changes it so that the user has to instantiate an instance
of the BPMPServer and connect it to the ClockServer.

Also, this commit also adds a platform specific initialisation function
to allow platform specific initialisation procedures to be carried out
before initialising the clock drivers. For the tx2, this involves
creating a BPMP interface using the BPMPServer server library and
registering it with the interface registration service so that clock
driver can find it and use it.