Ethdriver,tx2: Add interfaces for Reset,Clock,GPIO
Now that there are CAmkES components for the reset, clock, and GPIO
hardware, instead of initialising and handling all the related hardware
inside the Ethdriver component, we can instantiate these hardware
components and connect to them. This commit adds these changes so that
the Ethdriver now communicates to the these individual 'hardware'
components.

Also add server interface initialising and registering code so that the
Ethernet driver is able to find these interfaces and use them to
communicate with the 'hardware' components.