BenchUtiliz: Move idle thread config internal
The bench_to interface is used as the idle thread which needs to have
the lowest priority and is connected back in on itself to be able to
start the idle thread from the control thread. Make this connection
internal as it is not expected to be changed when integrated into a
project.