BenchUtiliz: Tweak measurement control process
This changes the behavior of the component so that the idle thread is
running continuously and the start and stop procedures sample values of
the idle thread's state that are continuously changing. This allows the
start and stop procedures to run on a different thread with a higher
priority and reduce the chance of measurement errors that can arise when
the idle thread gets a command to start executing, but it isn't
scheduled until partway through the measurement and records late start
times and likewise for finishing. The consequence of this change is that
when a measurement isn't being taken the idle thread will still run when
the system would otherwise be idle. We assume that a system using this
component is purpose configured for measuring idle time and doesn't care
about minimizing execution when measurements aren't being taken.