single-threaded: Add event handling tracing
This tracing uses the BenchUtiliz component to start and stop tracing of
event handlers within a component.  This can be turned on on a per
component instance by setting the configuration option enable_tracing=1.
The number of tracepoints is statically configured and can be set with
${instance_name}.number_tracepoints = 10. Where 10 is the default. If
more callback handlers are registered than what is statically
configured, tracing wont be performed for any additional handlers.