single-threaded: Merge last signal call with wait
For a virtqueue signal operation, combine the last signal with waiting
on the components endpoint. This is to optimize the case where a
component finished doing some work, signals another component and goes
back to sleep to remove an additional kernel entry/exit.

Note that this optimization is on by default and can be toggled off with
a configuration option. To turn this off, set the CAmkES attribute
"<interface_name>.disable_optimizations = 1", where interface name is
the name of the interface that's connected with another component with
the seL4VirtQueues connector.