templates: generated header files for connectors
This commit adds FROM_HEADER and TO_HEADER information for global
connectors missing this information. This resolves the issue of having
to declare the generated connector functions manually in user code.
An example, where this was done properly before is seL4GPIOServer.
The changes were also done in a way, that header information are
generated recursively if other connectors are reused, example being
get_notification.template.c/h in seL4Ethdriver.
Lastly global-connectors.cmake was adapted to include a foreach loop
for those connectors providing templates with FROM, FROM_HEADER, TO
and TO_HEADER information.

Signed-off-by: Thomas Wittal <thomas.wittal@hensoldt-cyber.de>