SerialServer: allow clients to be non-sequential
This commit adds a helper function to get a client structure based on
the badge. The allows for clients to be numerical, but not sequential,
and guarantees serial input will always align to the correct client on
each build.

A side-effect of this method is that the clients are switched to by
typing the ESCAPE_CHAR + badge number. This limits the number of clients
to single-digits, 0 -> 9.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>