SerialServer: initialize putchar interfaces
A zero badge is technically an unbadged notification, which would allow
the owner to distribute the capability. CAmkES makes sure the attributes
aren't set to zero, but that meant that "Client 0" was unused.

This commit initializes the putchar interfaces, and subtracts 1 from the
badge to get a client ID. This adds a runtime check to ensure a badge
of 0 cannot be used by the serial server.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>