Ethdriver: Add workaround for DMA cachability conf
Previously, the camkes runtime would ignore the cacheability attribute
of the DMA memory request and just return a static DMA memory mapping
that would work for the platform (cached on x86 and uncached on
non-x86). Now that the runtime respects the cacheability argument the
ethdriver component can just try both combinations and expect one of the
allocations to work if the component has been configured with a DMA
pool.

Signed-off-by: Kent McLeod <kent@kry10.com>